Add files via upload
main_w_DE.py is the assembly of Security Table generation (initial main.py) with NESS decision engine. The table is generated with inclusion of malicious nodes and nodes not checked through CA and NESS decision engine analyzes the security status for all nodes in the table and reports results.
the krb and kfb files contain decision rules.